requirements.txt
streamlit==1.18.1
tensorflow==2.17.0
numpy
opencv-python
pillow
